+++ src/gui/dialogs/history_server_dialog.cpp
@@ -141,7 +141,7 @@ void ServerHistoryDialog::refreshInfoFields(int index) {
 
   std::vector< std::vector<core::Field> > fields = infoFieldsFromType(server_->type());
   std::vector<core::Field> field = fields[index];
-  for (size_t i = 0; i < field.size(); ++i) {
+  for (int i = 0; i < field.size(); ++i) {
     core::Field fl = field[i];
     if (fl.isIntegral()) {
       serverInfoFields_->addItem(common::convertFromString<QString>(fl.name), i);
